requirements.txt
asgiref==3.5.2
black==22.3.0
click==8.1.3
Django==4.0.5
django-environ==0.8.1
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
install==1.3.5
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.1.1
platformdirs==2.5.2
PyJWT==2.4.0
pymemcache==3.5.2
python-dateutil==2.8.2
python-memcached==1.59
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
